Split competition code
