+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/noncompetition_code/AutonomousDemo.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.noncompetition_code;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
