+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/noncompetition_code/GamepadControl.java
@@ -1,9 +1,9 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.noncompetition_code;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.teamcode.prefs.BaseType;
+import org.firstinspires.ftc.teamcode.noncompetition_code.prefs.BaseType;
 
 @TeleOp(name = "Gamepad Control", group = "Control")
 public class GamepadControl extends LinearOpMode {
