+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/noncompetition_code/prefs/BaseType.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.prefs;
+package org.firstinspires.ftc.teamcode.noncompetition_code.prefs;
 
 public enum BaseType {
     twoWheeler,
