+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/noncompetition_code/utils/Offset.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.utils;
+package org.firstinspires.ftc.teamcode.noncompetition_code.utils;
 
 class Offset {
     private final double dx, dy;
